evaluation/human_evaluation/requirements.txt
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
streamlit==1.35.0
gspread==6.1.2
